Organizing files under higherkindness.skeuomorph
